requirements.txt
# Vector store
faiss-cpu>=1.8.0

# HTTP client (used by Ollama provider)
httpx>=0.27.0

# Data
pandas>=2.2.0
numpy>=1.26.0
pydantic>=2.7.0

# API
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# UI
streamlit>=1.35.0

# ML — structured model layer
xgboost>=2.0.0
shap>=0.45.0
scikit-learn>=1.4.0

# Experiment tracking
mlflow>=2.13.0

# Utilities
python-dotenv>=1.0.0
tenacity>=8.3.0
tiktoken>=0.7.0
tqdm>=4.66.0

# LLM providers — install whichever you use
# Groq + OpenAI both use the openai package
openai>=1.30.0           # required for Groq and OpenAI

# Optional providers (uncomment if needed)
# anthropic>=0.25.0      # Anthropic Claude

# Testing
pytest>=8.2.0
pytest-asyncio>=0.23.0
